Begin work on login provider to allow manual button to toggle the login screen
