+++ frontend/src/components/AAC/ManualModeButton.tsx
@@ -1,6 +1,8 @@
 import React, { useCallback } from "react";
 import Tile from "./Tile";
 import { useManualModeModelContext } from "@/react-state-management/providers/ManualModalProvider";
+import { useLoginProviderContext } from "@/react-state-management/providers/LoginPopupProvider";
+
 import { LongPressEventType, useLongPress } from "use-long-press";
 
 export const ManualBtnTestIds = {
@@ -9,26 +11,30 @@ export const ManualBtnTestIds = {
 
 export default function ManualModeButton() {
     const [isOpen, toggleModal] = useManualModeModelContext();
+    const [isLoginOpen, setLoginOpen] = useLoginProviderContext();
+
+    const toggleLoginOpen = () => setLoginOpen((prev) => !prev);
 
     const toggleModelHandler = () => {
-        console.log("Onclick triggered");
         toggleModal();
     }
 
-    const callback = useCallback(() => {
-        console.log("Long Pressed!")
-      }, []);
+    // const callback = useCallback(() => {
+    //     console.log(isLoginOpen);
+    //   }, []);
 
-    const bind = useLongPress(callback, {
-    onStart: () => console.log('Press started'),
-    onFinish: () => console.log('Long press finished'),
-    onCancel: toggleModelHandler,
-    onMove: () => console.log('Detected mouse or touch movement'),
-    filterEvents: () => true, // All events can potentially trigger long press (same as 'undefined')
-    threshold: 800, // In milliseconds
-    captureEvent: true, // Event won't get cleared after React finish processing it
-    cancelOnMovement: false, // Square side size (in pixels) inside which movement won't cancel long press
-    cancelOutsideElement: true, // Cancel long press when moved mouse / pointer outside element while pressing
+    const bind = useLongPress(() => {}, {
+        onFinish: () => {
+            console.log(isLoginOpen);
+            toggleLoginOpen();
+            console.log(isLoginOpen);
+        },
+        onCancel: toggleModelHandler,
+        filterEvents: () => true, // All events can potentially trigger long press (same as 'undefined')
+        threshold: 800, // In milliseconds
+        captureEvent: true, // Event won't get cleared after React finish processing it
+        cancelOnMovement: false, // Square side size (in pixels) inside which movement won't cancel long press
+        cancelOutsideElement: false, // Cancel long press when moved mouse / pointer outside element while pressing
     });
 
     return (
