+++ frontend/src/pages/index.tsx
@@ -6,6 +6,7 @@ import Canvas from "@/components/AAC/Canvas";
 import SuggestedTiles from "@/components/AAC/SuggestedTile";
 import RecentlyClickedTiles from "@/components/AAC/RecentlyClickedTiles";
 import CaretakerPopup from "@/components/AAC/CaretakerPopup";
+import Login from "@/components/Login";
 
 import RekognitionProvider from "@/react-state-management/providers/useRekognition";
 import TileProvider from "@/react-state-management/providers/tileProvider";
@@ -18,7 +19,7 @@ import SuggestedTilesProvider from "@/react-state-management/providers/Suggested
 import SimilarityProvider from "@/react-state-management/providers/useSimilarity";
 import InferenceProvider from "@/react-state-management/providers/InferenceProvider";
 import CaretakerPopupProvider from "@/react-state-management/providers/CaretakerPopupProvider";
-
+import LoginPopupProvider  from "@/react-state-management/providers/LoginPopupProvider";
 
 import useDisableZoom from "@/react-helpers/hooks/useDisableZoom";
 /**
@@ -26,7 +27,7 @@ import useDisableZoom from "@/react-helpers/hooks/useDisableZoom";
  * @returns the homepage for this app
  */
 export default function Home() {
-    
+
     useDisableZoom();
 
     return (
@@ -39,21 +40,24 @@ export default function Home() {
                                 <SuggestedTilesProvider>
                                     <UtteredTilesProvider>
                                         <ModalProvider>
-                                                    <div className = "static">
-                                                        <CaretakerPopupProvider>
-                                                            <CaretakerPopup />
-                                                        </CaretakerPopupProvider>
-                                                        <SelectedTilesActionBar />
-                                                        <ManualTilesPopup />
-                                                    </div>
-                                                    <div className="flex gap-2 w-full h-96 shrink h-calc-vh">
-                                                        <Canvas />
-                                                        <RecentlyClickedTiles />
-                                                    </div>
-                                                    <div className="flex gap-2 m-3 mt-1 justify-between items-center">
-                                                        <SuggestedTiles />
-                                                        <ManualModeButton />
-                                                    </div>
+                                            <LoginPopupProvider>
+                                                <div className="static">
+                                                    <CaretakerPopupProvider>
+                                                        <CaretakerPopup />
+                                                    </CaretakerPopupProvider>
+                                                    <SelectedTilesActionBar />
+                                                    <ManualTilesPopup />
+                                                </div>
+                                                <div className="flex gap-2 w-full h-96 shrink h-calc-vh">
+                                                    <Canvas />
+                                                    <RecentlyClickedTiles />
+                                                </div>
+                                                <div className="flex gap-2 m-3 mt-1 justify-between items-center">
+                                                    <SuggestedTiles />
+                                                    <ManualModeButton />
+                                                    <Login />
+                                                </div>
+                                            </LoginPopupProvider>
                                         </ModalProvider>
                                     </UtteredTilesProvider>
                                 </SuggestedTilesProvider>
